Move example to  new directory. Remove RN project leftovers from root. npmignore example
