+++ SwipeListExample/.watchmanconfig
