+++ SwipeListExample/android/app/src/main/res/mipmap-xhdpi/ic_launcher.png
