+++ SwipeListExample/android/app/src/main/res/values/styles.xml
