+++ SwipeListExample/android/build.gradle
