+++ SwipeListExample/android/gradle.properties
