+++ SwipeListExample/android/gradle/wrapper/gradle-wrapper.jar
