+++ SwipeListExample/android/gradle/wrapper/gradle-wrapper.properties
