+++ SwipeListExample/android/gradlew
