+++ SwipeListExample/android/gradlew.bat
