+++ SwipeListExample/android/keystores/debug.keystore.properties
@@ -1,0 +1,4 @@
+key.store=debug.keystore
+key.alias=androiddebugkey
+key.store.password=android
+key.alias.password=android
