+++ SwipeListExample/android/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'SwipeListExample'
+
+include ':app'
