+++ SwipeListExample/index.android.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import App from './example';
+
+AppRegistry.registerComponent('SwipeListExample', () => App);
