+++ SwipeListExample/ios/SwipeListExample/AppDelegate.h
