+++ SwipeListExample/ios/SwipeListExample/Images.xcassets/AppIcon.appiconset/Contents.json
