+++ SwipeListExample/ios/SwipeListExample/Info.plist
