+++ SwipeListExample/ios/SwipeListExampleTests/Info.plist
