+++ SwipeListExample/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "SwipeListExample",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "start": "node node_modules/react-native/local-cli/cli.js start"
+  },
+  "dependencies": {
+    "react": "^15.0.2",
+    "react-native": "^0.26.2",
+    "react-native-swipe-list-view": "file:../"
+  }
+}
