--- index.android.js
@@ -1,4 +1,0 @@
-import { AppRegistry } from 'react-native';
-import App from './example';
-
-AppRegistry.registerComponent('reactNativeSwipeList', () => App);
